+++ docs/changelog.md
@@ -4,6 +4,8 @@
 
 - The terminology for relationships can now be customised.
 - Added support for icons on element styles.
+- Top-level deployment nodes can now be given an environment property, to represent which deployment environment they belong to (e.g. "Development", "Live", etc).
+- Relationships can no longer be created between container instances (__breaking change__).
 - Fixed issue #40 (WorkspaceUtils does not load containers' properties).
 
 ## 0.9.0 (8th November 2018)